Make it possible to set a custom user and cache providers when using legacy store
Closes #15008
