+++ quarkus/tests/integration/src/test-providers/resources/com/acme/provider/legacy/jpa/entity/persistence.xml
@@ -22,7 +22,7 @@
         http://java.sun.com/xml/ns/persistence
         http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
     <persistence-unit name="user-store" transaction-type="JTA">
-        <class>com.acme.provider.user.Realm</class>
+        <class>com.acme.provider.legacy.jpa.entity.Realm</class>
         <properties>
             <property name="hibernate.dialect" value="io.quarkus.hibernate.orm.runtime.dialect.QuarkusH2Dialect" />
             <!-- Sets the name of the datasource to be the same as the datasource name in quarkus.properties-->
