+++ quarkus/tests/integration/src/test-providers/resources/com/acme/provider/legacy/jpa/entity/quarkus.properties
