+++ quarkus/tests/integration/src/test-providers/resources/com/acme/provider/legacy/jpa/user/org.keycloak.models.UserProviderFactory
@@ -1,0 +1,1 @@
+com.acme.provider.legacy.jpa.user.MyUserProviderFactory
